% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rain.R
\name{calc_rain}
\alias{calc_rain}
\title{Average Percipitation 1995-2015}
\usage{
calc_rain(rain_data)
}
\arguments{
\item{rain_data}{is from data within this package that has the average rainfall between 1995-2015}
}
\value{
rain_matrix (max and min rain years and their average)
}
\description{
This is a function that determines the wet and dry year between 1995 and 2015
}
\author{
Sidney
}
