% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_temp.R
\name{calc_temp}
\alias{calc_temp}
\title{Average Temperature 1995-2015}
\usage{
calc_temp(temp_data)
}
\arguments{
\item{temp_data}{is from data within this package that has the average temperature between 1995-2015}
}
\value{
rain_matrix (max and min rain years and their average)
}
\description{
This is a function that determines the cold and hot years between 1995 and 2015
}
\author{
Sidney
}
