% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kbdi_index.R
\name{kbdi_index}
\alias{kbdi_index}
\title{KBDI}
\usage{
kbdi_index(Temperature, Rain, MAP)
}
\arguments{
\item{Temperature}{The daily temperature (Degrees F)}

\item{Rain}{The daily rain fall amoung (mm)}

\item{MAP}{Mean Annual Precipitation (mm)}
}
\value{
kbdi
}
\description{
This is a function that calcualtes the Keetch-Byram Drought Index. This index can be used in fire preparedness. Low numbers indicate low fire risk, while high numbers indicate high fire risk.
}
\author{
Sidney
}
