% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rain_data.R
\docType{data}
\name{rain_data}
\alias{rain_data}
\title{Rain Data from Sidney}
\format{A data frame with 20 rows and 3 columns
\describe{
\item{year}{year}
\item{avg_precip}{average amount of rain that year (inches)}
\item{anomaly}{how different that year was from the average}
}}
\usage{
rain_data
}
\description{
Data from how much average rain existed from 1995-2015
}
\keyword{datasets}
