% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_data.R
\docType{data}
\name{temp_data}
\alias{temp_data}
\title{Temperature Data from Sidney}
\format{A data frame with 20 rows and 3 columns
\describe{
\item{year}{year}
\item{avg_temp}{average temperature that year (degrees Fahrenheit)}
\item{temp_anomaly}{how different that year was from the average}
}}
\usage{
temp_data
}
\description{
Data from what the average temperature was from 1995-2015
}
\keyword{datasets}
